data-preprocessing/requirements.txt
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.1
clean-text==0.4.0
cleantext==1.1.4
emoji==1.7
regex==2023.6.3
beautifulsoup4==4.12.2